5. Project Executable Files/Flask/requirements.txt
flask
tensorflow
numpy
Pillow
werkzeug
streamlit
